requirements.txt
matplotlib
numpy>=1.12.0
opencv-python
pandas
pillow
protobuf==3.2.0
scipy
sklearn
